+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Update `get` function to accept a key path
+* Remove `getIn` function
 
 ## 3.0.2 (2019-04-09)